Bug 57915: Fix Dev2Hex for numbers larger than Integer.MAX_VALUE and less than Integer.MIN_VALUE
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1703672 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/formula/functions/Dec2Hex.java
@@ -112,7 +112,7 @@ public final class Dec2Hex extends Var1or2ArgFunction implements FreeRefFunction
             hex = String.format(Locale.ROOT, "%0"+placesNumber+"X", number1.intValue());
         }
         else {
-            hex = Integer.toHexString(number1.intValue());
+            hex = Long.toHexString(number1.longValue());
         }
 
         if (number1 < 0) {
